Bug fixes and updates
Update to the latest version of BirdbrainBLE. Add a reference to the robot manager delegate to compensate for the package moving to weak references for delegates. Add missing sensor request to backendServer. Cleanup fronendServer and fix 2 bugs (ble video was sometimes playing after screen sleep and auto reconnect was not working for devices that had previously been disconnected by the user). Update credits that appear in "About BlueBird Connector". Fix battery monitor bug (red level was not being displayed).